legal-rag/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.1
cut-cross-entropy==25.1.1
datasets==3.5.0
dill==0.3.8
docstring_parser==0.16
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
hf_transfer==0.1.9
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==11.2.1
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==19.0.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
sentencepiece==0.2.0
shtab==1.7.2
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1
tqdm==4.67.1
transformers==4.51.3
triton==3.1.0
trl==0.15.2
typeguard==4.4.2
typing_extensions==4.13.2
tyro==0.9.19
tzdata==2025.2
unsloth @ git+https://github.com/unslothai/unsloth.git@7a8f99e1890213cdd01a3ab6c3e13174a96e8220
unsloth_zoo==2025.4.1
urllib3==2.4.0
xformers==0.0.29.post1
xxhash==3.5.0
yarl==1.20.0
